Added command line option for starting clients on a uniform interval
With the option -startup_interval the client will, spread the client startups
evenly out on an interval of time designated by this configuration option. This
helps in scale testing, as all clients will not come online at the same time,
and hence bring the server to its knees.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>