MEN-2879: Document the mender-stress-test-client
Updated the Readme to give a better overview of the quirks and considerations
when running the mender-stress-test-client. This is added in three chapters,
getting started, one for the client itself (Working with the client), and one
for working with the demo server, aptly named (Working with the demo server).

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>